requirements.txt
# Deep learning
torch>=2.0.0
transformers>=4.38.0
accelerate>=0.27.0

# NLP
spacy>=3.7.0
# Run after install: python -m spacy download en_core_web_sm

# Data
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# RSS fetching
feedparser>=6.0.11
python-dateutil>=2.8.2

# Database
pymongo>=4.7.0
certifi>=2024.0.0

# API
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# Config
python-dotenv>=1.0.1
